Simple GITGraph class to represent a directed acyclic graph (DAG)
The graph class is a wrapper around an NSMutableDictionary of node objects.
Each dictionary entry consists of a sha1 key (40-char NSString) and a
GITObjectNode value that wraps the corresponding GITObject (GITCommit for
now). Because of this arrangement, the graph acts like a standard Foundation
collection in the sense that it retains objects that are added, and released
objects upon removal.

In addition to methods for adding/removing and querying nodes, the graph also
supports specialized sort methods that return NSArrays of nodes sorted by date
or topology. These lists are identical to the lists of objects obtained from
git-rev-list with the default, --topo-order or --date-order flags.

The interface currently lacks much documentation, but is relatively
straightforward. For use examples, see the unit tests or the source
'git-rev-list' tool.

NOTES
-----
Each of the 'sort' utilities requires the entire graph to be built and
held in memory prior to sorting. I think this is required for the topological
sort, but not for the date-based sort. I think it should be possible to
abstract out some of the functionality into enumerators or generators that can
spit out objects on demand, without requiring as much overhead.

Signed-off-by: Geoff Garside <geoff@geoffgarside.co.uk>

Signed-off-by: Benjamin Stiglitz <ben@tanjero.com>