Add GITNode class to represent a GITObject in a graph.
A GITNode is a simple wrapper class that store a pointer to a GITObject and
also caches/stores information related to the graph structure. A GITObject
(currently a GITCommit) is required to create a node, since the nodes key is
derived from the object's sha1. This key is used to uniquely identify the node
in a graph.

The idea behind separating the actual objects from the graph representation
is:

1. convenience of not added extra properties to GITObject/GITCommit
2. encapsulation of graph storage/traversal algorithms
3. possibility to create thread-safe graph structure for multi-threaded
processing.

Once constructed, we can still do traversals and sorting operations without
the objects. They could be released to save space. In addition, this also
makes it possible to create a minimal copy of the graph for multi-threaded
operations, should the need arise. On the other hand, it may turn out that the
overhead of creating extra objects is a bad idea.

I'm also aware that this is a fairly naive way to handle a graph. There are
almost certainly better/more efficient ways of storing and accessing the same
information. This seems to work for now, but is subject to change should the
need arise.

Rough estimate of memory overhead for a graph of the git repository starting
from HEAD, containing ~18400 commits:

graph + nodes (with objects) = 23 MB
graph + nodes (objects released) = 8 MB

Signed-off-by: Geoff Garside <geoff@geoffgarside.co.uk>

Signed-off-by: Benjamin Stiglitz <ben@tanjero.com>