git-rev-list CLI tool: uses API to reproduce output from 'git rev-list'
The tool currently supports output identical[1] to the following commands:

git rev-list <commit-id>
git rev-list --date-order <commit-id>
git rev-list --topo-order <commit-id>

Note that listing multiple commits, ranges or excluding commits is not
supported, but branch heads can be specified in place of <commit-id>. See use
for more info.

This tool is exceptionally useful for profiling and comparison with the real
git-rev-list. It plays well with Shark and Instruments as long as you setup
the command arguments properly (hint: see the "Executables" grouping in
Xcode). The '-r' flag allows you to specify any repo, so you can clone out the
linux kernel and use that for testing, ect..

[1] This is based on my testing with the git repository, comparing output
using HEAD as the commit with the options specified above. This was actually a
PITA to sort out, especially since --date-order is almost always identical to
the default ordering, except in some rare cases. The test repo now contains a
small version of a branch with enough strange date-madness edge-cases that
will produce a unique sort ordering in each case. This should be useful for
further development of these routines.

Signed-off-by: Geoff Garside <geoff@geoffgarside.co.uk>

Signed-off-by: Benjamin Stiglitz <ben@tanjero.com>